+++ src/app/class/[id]/page.tsx
@@ -5,8 +5,8 @@ import Review from '@/components/Review/Review';
 import ReviewComment from '@/components/Review/ReviewComment';
 import Nav from '@/components/Nav/Nav';
 import Like from '@/components/Like/Like';
+import Notice from '@/components/ClassNotice/Notice';
 import Apply from './_components/Apply';
-import Notice from './_components/Notice';
 import ProfileButtons from './_components/ProfileButtons';
 import { CLASS_SECTIONS } from '@/constants/constants';
 import { dummyClass, dummyImgURL } from '@/constants/dummy';
