+++ src/components/ClassNotice/Notice.tsx
